requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
sqlalchemy
passlib
pyjwt
bcrypt